requirements.txt
viam-sdk~=0.32.0
dbus-python~=1.3.2
PyGObject~=3.50.0 
pycairo~=1.27.0